+++ cmd/myshell/builtin/builtins.go
@@ -1,6 +1,6 @@
 package builtin
 
-var builtins = []string{"exit", "echo", "type", "cls", "pwd", "cd", "thx", "dirs", "help", "sleep", "mkdir", "ls", "rm", "mv", "time"}
+var builtins = []string{"exit", "echo", "type", "cls", "pwd", "cd", "thx", "dirs", "help", "sleep", "mkdir", "ls", "rm", "mv", "time", "touch"}
 
 func GetBuiltins() []string {
 	return builtins
@@ -38,9 +38,9 @@ func HandleBuiltins(command string, args []string) bool {
 	case "mv":
 		Mv(args)
 	case "time":
-		//fmt.Print("time: not implemented yet")
 		Time(args)
-
+	case "touch":
+		Touch(args)
 	default:
 		is_builtin = false
 	}